Use `underscore_name` instead of `name.underscore` (#1094)
There's a bug in the trix field at the moment, where the `input` attribute of the `<trix-editor>` tag can include spaces whenever you have a resource with more than one word in it's name. If you have a resource named `MyPost` then `@resource.name` will be `My post`, `@resource.name.underscore` will be `my post` and therefore `trix_id` will be `trix_my post_foo` (if the field is named `foo`).

To fix this I've changed it to use the class name instead of the human readable name,
then calling underscore on that.